dedupe-tile-component * Use Tile and TileHeader in ChronoBikeTile * Use css-modules for ChronoBikeTile and ChronoBikeTileRow
