--- src/dashboards/Chrono/BikeTile/BikeTile.scss
@@ -1,43 +1,0 @@
-@use '../../../assets/styles/Variables.scss';
-
-.chrono .bikerow {
-    display: flex;
-    border-bottom: 2px solid var(--tavla-border-color);
-    padding: 1rem 0;
-
-    &__label {
-        margin: 0;
-        margin-bottom: 0.25rem;
-        color: var(--tavla-font-color);
-    }
-
-    &__texts {
-        flex-direction: column;
-        flex-grow: 1;
-    }
-
-    &__icon {
-        min-width: 2rem;
-        margin-right: 1rem;
-        font-size: Variables.$font-sizes-extra-large4;
-    }
-
-    &__walking-time {
-        font-style: normal;
-        font-weight: normal;
-        font-size: 1rem;
-        margin: 0.5rem 0;
-        color: var(--tavla-label-font-color);
-        overflow: hidden;
-        white-space: nowrap;
-        text-overflow: ellipsis;
-    }
-
-    &__sublabels {
-        display: grid;
-        grid-template-columns: repeat(3, 1fr);
-        font-size: 1.25rem;
-        justify-content: space-between;
-        max-width: 18rem;
-    }
-}
