+++ src/dashboards/Chrono/ChronoBikeTile/ChronoBikeTile.module.scss
@@ -1,0 +1,3 @@
+.chronoBikeTile {
+    height: 100%;
+}
