+++ src/dashboards/Chrono/ChronoBikeTile/ChronoBikeTile.tsx
@@ -1,21 +1,22 @@
 import React, { useState, useEffect, useMemo } from 'react'
 import { colors } from '@entur/tokens'
 import { BicycleIcon } from '@entur/icons'
-import { Tile } from '../components/Tile/Tile'
+import { Tile } from '../../../components/Tile/Tile'
+import { TileHeader } from '../../../components/TileHeader/TileHeader'
 import { useSettings } from '../../../settings/SettingsProvider'
 import { IconColorType } from '../../../types'
 import { getTranslation } from '../../../utils/utils'
 import { getIconColorType } from '../../../utils/icon'
 import { useWalkInfo, WalkInfo } from '../../../logic/use-walk-info/useWalkInfo'
 import { UseRentalStations_StationFragment } from '../../../../graphql-generated/mobility-v2'
-import { TileRow } from '../components/TileRow/TileRow'
-import './BikeTile.scss'
+import { ChronoBikeTileRow } from '../ChronoBikeTileRow/ChronoBikeTileRow'
+import classes from './ChronoBikeTile.module.scss'
 
 function getWalkInfo(walkInfos: WalkInfo[], id: string): WalkInfo | undefined {
     return walkInfos.find((walkInfo) => walkInfo.stopId === id)
 }
 
-const BikeTile = ({ stations }: Props): JSX.Element => {
+const ChronoBikeTile = ({ stations }: Props): JSX.Element => {
     const [settings] = useSettings()
     const [iconColorType, setIconColorType] = useState<IconColorType>(
         IconColorType.CONTRAST,
@@ -42,17 +43,18 @@ const BikeTile = ({ stations }: Props): JSX.Element => {
     const walkInfo = useWalkInfo(stationDestinations)
 
     return (
-        <Tile
-            title="Bysykkel"
-            icons={[
-                <BicycleIcon
-                    key="bike-tile-icon"
-                    color={colors.transport[iconColorType].mobility}
-                />,
-            ]}
-        >
+        <Tile className={classes.chronoBikeTile}>
+            <TileHeader
+                title="Bysykkel"
+                icons={[
+                    <BicycleIcon
+                        key="bike-tile-icon"
+                        color={colors.transport[iconColorType].mobility}
+                    />,
+                ]}
+            />
             {stations.map((station) => (
-                <TileRow
+                <ChronoBikeTileRow
                     key={station.id}
                     icon={
                         <BicycleIcon
@@ -91,4 +93,4 @@ interface Props {
     stations: UseRentalStations_StationFragment[]
 }
 
-export { BikeTile }
+export { ChronoBikeTile }
