--- src/dashboards/Chrono/components/TileRow/TileRow.scss
@@ -1,53 +1,0 @@
-@use '../../../../assets/styles/Variables.scss';
-
-.chrono .tilerow {
-    display: flex;
-    border-bottom: 2px solid var(--tavla-border-color);
-    padding: 1rem 0;
-    overflow-x: hidden;
-    overflow-y: hidden;
-
-    &:last-child {
-        border-bottom: 0;
-        padding-bottom: 0;
-    }
-
-    &__label {
-        margin: 0;
-        margin-bottom: 0.5rem;
-        color: var(--tavla-font-color);
-    }
-
-    &__icon {
-        min-width: 2rem;
-        margin-right: 1rem;
-        font-size: Variables.$font-sizes-extra-large4;
-    }
-
-    &__texts {
-        flex-direction: column;
-        flex-grow: 1;
-        overflow: hidden;
-        white-space: nowrap;
-        text-overflow: ellipsis;
-    }
-
-    &__walking-time {
-        font-style: normal;
-        font-weight: normal;
-        font-size: 1rem;
-        margin: 0.5rem 0;
-        color: var(--tavla-label-font-color);
-        overflow: hidden;
-        white-space: nowrap;
-        text-overflow: ellipsis;
-    }
-
-    &__sublabels {
-        display: grid;
-        grid-template-columns: repeat(20, 1fr);
-        justify-content: space-between;
-        min-width: 130rem;
-        font-size: Variables.$font-sizes-extra-large;
-    }
-}
